WC Post Types: Move shortcodes to deprecated file
